+++ R/driver_dbi.R
@@ -3,13 +3,12 @@
 ##' (though practically this works only for SQLite and Postgres until
 ##' some MySQL dialect translation is done).  To connect, you must
 ##' provide the \emph{driver} object (e.g., \code{RSQLite::SQLite()},
-##' \code{RPostgreSQL::PostgreSQL()}, \code{RPostgres::Postgres()} as
-##' the first argument,
+##' or \code{RPostgres::Postgres()} as the first argument.
 ##'
-##' Because the DBI package
-##' specifies a uniform interface for the using DBI compliant
-##' databases, you need only to provide a connection object.  storr
-##' does not do anything to help create the connection object itself.
+##' Because the DBI package specifies a uniform interface for the
+##' using DBI compliant databases, you need only to provide a
+##' connection object.  storr does not do anything to help create the
+##' connection object itself.
 ##'
 ##' The DBI storr driver works by using two tables; one mapping keys
 ##' to hashes, and one mapping hashes to values.  Two table names need
@@ -103,12 +102,14 @@ storr_dbi <- function(tbl_data, tbl_keys, con, args = NULL, binary = NULL,
         default_namespace)
 }
 
+
 ##' @rdname storr_dbi
 ##' @export
 driver_dbi <- function(tbl_data, tbl_keys, con, args = NULL, binary = NULL,
                        hash_algorithm = NULL) {
   R6_driver_DBI$new(tbl_data, tbl_keys, con, args, binary, hash_algorithm)
 }
+
 
 ## SQL interpolation
 ##
@@ -417,6 +418,7 @@ R6_driver_DBI <- R6::R6Class(
     }
   ))
 
+
 dbi_supports_binary <- function(con) {
   ## Very little binary support exists; requires newfangled DBI and
   ## new RSQLite.  None of the other connection types supports binary
@@ -429,6 +431,7 @@ dbi_supports_binary <- function(con) {
   }
   supports_binary
 }
+
 
 dbi_use_binary <- function(con, tbl_data, binary) {
   supports_binary <- dbi_supports_binary(con)
@@ -462,8 +465,10 @@ dbi_use_binary <- function(con, tbl_data, binary) {
   }
 }
 
+
 ## 15 'f's - no hash algo has this length
 STORR_DBI_CONFIG_HASH <- paste(rep("f", 15), collapse = "")
+
 
 ## This is going to hold prepared queries for the core bits of the
 ## SQL; these vary fairly wildly by dialect, and some care is needed
@@ -565,6 +570,7 @@ driver_dbi_sql_compat <- function(dialect, tbl_data, tbl_keys) {
   ret
 }
 
+
 driver_dbi_mkey_prepare <- function(key, namespace, placeholder) {
   nk <- join_key_namespace(key, namespace)
   if (nk$n == 0L) {
@@ -617,22 +623,24 @@ driver_dbi_mkey_prepare <- function(key, namespace, placeholder) {
        values = values)
 }
 
+
 driver_dbi_mhash_prepare <- function(hash, placeholder) {
   paste(sprintf(placeholder, seq_along(hash)), collapse = ", ")
 }
+
 
 driver_dbi_dialect <- function(con) {
   if (inherits(con, "SQLiteConnection")) {
     "sqlite"
   } else if (inherits(con, c("PqConnection", "PostgreSQLConnection"))) {
-    ## Before 0.9.5 there was no simple way of implementing the
+    ## Before 9.5.0 there was no simple way of implementing the
     ## "INSERT OR REPLACE INTO" pattern (via INSERT INTO ... ON
     ## CONFLICT REPLACE" and I'm just going to require a recent
     ## version for simplicity.
     v <- pg_server_version(con)
-    if (v < numeric_version("0.9.5")) {
+    if (v < numeric_version("9.5.0")) {
       stop(sprintf(
-        "Version %s of postgresql server is not supported (need >= 0.9.5)", v))
+        "Version %s of postgresql server is not supported (need >= 9.5.0)", v))
     }
     "postgresql"
   } else {
@@ -641,19 +649,39 @@ driver_dbi_dialect <- function(con) {
   }
 }
 
+
 driver_classes <- function() {
   c("SQLiteConnection",
     "PqConnection", "PostgreSQLConnection")
 }
+
 
 group_placeholders <- function(placeholder, n, times) {
   p <- matrix(sprintf(placeholder, seq_len(n * times)), n)
   paste(sprintf("(%s)", apply(p, 2, paste, collapse = ", ")), collapse = ", ")
 }
 
+
 pg_server_version <- function(con) {
-  numeric_version(DBI::dbGetQuery(con, "show server_version")[[1L]])
+  pg_server_version_parse(DBI::dbGetQuery(con, "show server_version_num")[[1L]])
 }
+
+
+pg_server_version_parse <- function(v) {
+  v <- as.integer(v)
+  major <- v %/% 10000
+  if (major >= 10) {
+    minor <- v %% 10000
+    str <- sprintf("%d.%d", major, minor)
+  } else {
+    v <- v %% 10000
+    minor <- v %/% 100
+    patch <- v %% 100
+    str <- sprintf("%d.%d.%d", major, minor, patch)
+  }
+  numeric_version(str)
+}
+
 
 dbi_connection_factory <- function(drv, args) {
   if (is.null(drv)) {
@@ -670,6 +698,7 @@ dbi_connection_factory <- function(drv, args) {
     }
   }
 }
+
 
 assert_valid_table_name <- function(x, name = deparse(substitute(x))) {
   assert_scalar_character(x, name)