+++ tests/testthat/test-auto.R
@@ -18,37 +18,28 @@ test_that("dbi (sqlite)", {
   }
 })
 
-pg_tester <- function(ctor) {
-  force(ctor)
-  function(dr = NULL, ...) {
-    if (is.null(dr)) {
-      prefix <- paste(sample(letters, 8), collapse = "")
-      dr <- list(con = DBI::dbConnect(ctor()),
-                 tbl_data = sprintf("storr_%s_data", prefix),
-                 tbl_keys = sprintf("storr_%s_keys", prefix))
-    }
-    driver_dbi(dr$tbl_data, dr$tbl_keys, dr$con, ...)
-  }
-}
-
 test_that("dbi (postgres via RPostgres)", {
   skip_on_cran()
   if (requireNamespace("RPostgres", quietly = TRUE)) {
-    if (has_postgres(RPostgres::Postgres)) {
-      storr::test_driver(pg_tester(RPostgres::Postgres))
+    if (has_postgres()) {
+      pg_create <- function(dr = NULL, ...) {
+        if (is.null(dr)) {
+          prefix <- paste(sample(letters, 8), collapse = "")
+          dr <- list(con = DBI::dbConnect(RPostgres::Postgres()),
+                     tbl_data = sprintf("storr_%s_data", prefix),
+                     tbl_keys = sprintf("storr_%s_keys", prefix))
+        }
+        driver_dbi(dr$tbl_data, dr$tbl_keys, dr$con, ...)
+      }
+      storr::test_driver(pg_create)
     }
   }
 })
 
-test_that("dbi (postgres via RPostgreSQL)", {
-  skip_on_cran()
-  if (requireNamespace("RPostgreSQL", quietly = TRUE)) {
-    oo <- options(warnPartialMatchArgs = FALSE)
-    if (!is.null(oo$warnPartialMatchArgs)) {
-      on.exit(options(oo))
-    }
-    if (has_postgres(RPostgreSQL::PostgreSQL)) {
-      storr::test_driver(pg_tester(RPostgreSQL::PostgreSQL))
-    }
-  }
+test_that("multistorr (env + rds)", {
+  .driver_create <-
+    storr::test_driver(function(dr = NULL, ...)
+      driver_multistorr(
+        driver_environment(dr$keys$envir),
+        driver_rds(dr$data$path %||% tempfile("storr_"), ...)))
 })